# Re-use priority queue in StaticRTree