Gives slight performance boost, because we don't have allocations in PQ...




<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1138.37<br/>plain u32: 1147.21<br/>aliased double: 1182.51<br/>plain double: 1177.4 | aliased u32: 1139.58<br/>plain u32: 1146.7<br/>aliased double: 1180<br/>plain double: 1180.56 |
| e2e_match_ch | Ops: 45.54 ± 0.08 ops/s. Best: 45.44 ops/s<br/>Total: 2876.63ms ± 5.03ms. Best: 2866.25ms<br/>Min time: 2.22ms ± 0.03ms<br/>Mean time: 21.96ms ± 0.04ms<br/>Median time: 15.59ms ± 0.15ms<br/>95th percentile: 72.49ms ± 0.46ms<br/>99th percentile: 86.37ms ± 0.41ms<br/>Max time: 92.40ms ± 0.89ms | Ops: 45.92 ± 0.05 ops/s. Best: 45.87 ops/s<br/>Total: 2852.55ms ± 3.01ms. Best: 2848.39ms<br/>Min time: 2.21ms ± 0.04ms<br/>Mean time: 21.78ms ± 0.02ms<br/>Median time: 15.41ms ± 0.08ms<br/>95th percentile: 71.44ms ± 0.20ms<br/>99th percentile: 85.63ms ± 0.35ms<br/>Max time: 91.40ms ± 0.32ms |
| e2e_match_mld | Ops: 62.57 ± 0.07 ops/s. Best: 62.42 ops/s<br/>Total: 2093.53ms ± 2.95ms. Best: 2088.50ms<br/>Min time: 1.81ms ± 0.03ms<br/>Mean time: 15.98ms ± 0.02ms<br/>Median time: 8.56ms ± 0.10ms<br/>95th percentile: 53.41ms ± 0.34ms<br/>99th percentile: 61.57ms ± 0.17ms<br/>Max time: 71.59ms ± 0.32ms | Ops: 62.73 ± 0.07 ops/s. Best: 62.64 ops/s<br/>Total: 2088.24ms ± 2.47ms. Best: 2083.35ms<br/>Min time: 1.82ms ± 0.02ms<br/>Mean time: 15.94ms ± 0.02ms<br/>Median time: 8.52ms ± 0.05ms<br/>95th percentile: 53.11ms ± 0.15ms<br/>99th percentile: 61.29ms ± 0.33ms<br/>Max time: 71.24ms ± 0.57ms |
| e2e_nearest_ch | Ops: 808.28 ± 5.28 ops/s. Best: 795.49 ops/s<br/>Total: 1237.19ms ± 8.38ms. Best: 1228.90ms<br/>Min time: 1.06ms ± 0.01ms<br/>Mean time: 1.24ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.64ms ± 0.01ms<br/>99th percentile: 1.69ms ± 0.02ms<br/>Max time: 4.42ms ± 2.68ms | Ops: 808.10 ± 4.31 ops/s. Best: 797.79 ops/s<br/>Total: 1237.48ms ± 7.01ms. Best: 1231.17ms<br/>Min time: 1.06ms ± 0.01ms<br/>Mean time: 1.24ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.64ms ± 0.01ms<br/>99th percentile: 1.69ms ± 0.01ms<br/>Max time: 4.07ms ± 2.34ms |
| e2e_nearest_mld | Ops: 798.94 ± 3.73 ops/s. Best: 790.94 ops/s<br/>Total: 1251.53ms ± 5.98ms. Best: 1244.49ms<br/>Min time: 1.07ms ± 0.00ms<br/>Mean time: 1.25ms ± 0.01ms<br/>Median time: 1.16ms ± 0.01ms<br/>95th percentile: 1.65ms ± 0.01ms<br/>99th percentile: 1.70ms ± 0.01ms<br/>Max time: 4.10ms ± 2.37ms | Ops: 810.22 ± 5.80 ops/s. Best: 796.31 ops/s<br/>Total: 1234.19ms ± 9.12ms. Best: 1225.97ms<br/>Min time: 1.05ms ± 0.01ms<br/>Mean time: 1.23ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.63ms ± 0.01ms<br/>99th percentile: 1.68ms ± 0.01ms<br/>Max time: 4.20ms ± 2.37ms |
| e2e_route_ch | Ops: 348.33 ± 1.68 ops/s. Best: 344.53 ops/s<br/>Total: 2870.57ms ± 14.28ms. Best: 2852.99ms<br/>Min time: 1.29ms ± 0.01ms<br/>Mean time: 2.87ms ± 0.01ms<br/>Median time: 2.89ms ± 0.02ms<br/>95th percentile: 3.78ms ± 0.03ms<br/>99th percentile: 4.21ms ± 0.04ms<br/>Max time: 6.61ms ± 1.78ms | Ops: 346.30 ± 0.89 ops/s. Best: 345.02 ops/s<br/>Total: 2887.93ms ± 7.67ms. Best: 2875.99ms<br/>Min time: 1.28ms ± 0.01ms<br/>Mean time: 2.89ms ± 0.01ms<br/>Median time: 2.91ms ± 0.01ms<br/>95th percentile: 3.82ms ± 0.04ms<br/>99th percentile: 4.21ms ± 0.06ms<br/>Max time: 6.50ms ± 1.72ms |
| e2e_route_mld | Ops: 291.67 ± 0.90 ops/s. Best: 289.54 ops/s<br/>Total: 3428.51ms ± 11.27ms. Best: 3418.43ms<br/>Min time: 1.27ms ± 0.02ms<br/>Mean time: 3.43ms ± 0.01ms<br/>Median time: 3.47ms ± 0.02ms<br/>95th percentile: 4.65ms ± 0.01ms<br/>99th percentile: 5.09ms ± 0.05ms<br/>Max time: 7.51ms ± 1.46ms | Ops: 289.23 ± 0.79 ops/s. Best: 288.12 ops/s<br/>Total: 3457.11ms ± 9.50ms. Best: 3440.72ms<br/>Min time: 1.30ms ± 0.02ms<br/>Mean time: 3.46ms ± 0.01ms<br/>Median time: 3.50ms ± 0.01ms<br/>95th percentile: 4.71ms ± 0.04ms<br/>99th percentile: 5.18ms ± 0.05ms<br/>Max time: 7.23ms ± 1.58ms |
| e2e_table_ch | Ops: 313.54 ± 1.33 ops/s. Best: 310.86 ops/s<br/>Total: 3189.46ms ± 12.83ms. Best: 3175.24ms<br/>Min time: 1.65ms ± 0.02ms<br/>Mean time: 3.19ms ± 0.01ms<br/>Median time: 3.19ms ± 0.01ms<br/>95th percentile: 4.39ms ± 0.03ms<br/>99th percentile: 4.71ms ± 0.06ms<br/>Max time: 7.20ms ± 2.34ms | Ops: 314.84 ± 0.99 ops/s. Best: 312.49 ops/s<br/>Total: 3176.12ms ± 10.49ms. Best: 3163.89ms<br/>Min time: 1.68ms ± 0.02ms<br/>Mean time: 3.18ms ± 0.01ms<br/>Median time: 3.17ms ± 0.01ms<br/>95th percentile: 4.35ms ± 0.01ms<br/>99th percentile: 4.70ms ± 0.06ms<br/>Max time: 7.16ms ± 2.31ms |
| e2e_table_mld | Ops: 110.61 ± 0.22 ops/s. Best: 110.20 ops/s<br/>Total: 9041.10ms ± 18.58ms. Best: 9018.57ms<br/>Min time: 3.67ms ± 0.03ms<br/>Mean time: 9.04ms ± 0.02ms<br/>Median time: 9.01ms ± 0.02ms<br/>95th percentile: 13.79ms ± 0.06ms<br/>99th percentile: 14.70ms ± 0.11ms<br/>Max time: 17.01ms ± 1.40ms | Ops: 110.51 ± 0.17 ops/s. Best: 110.10 ops/s<br/>Total: 9048.60ms ± 14.07ms. Best: 9033.62ms<br/>Min time: 3.65ms ± 0.02ms<br/>Mean time: 9.05ms ± 0.01ms<br/>Median time: 9.03ms ± 0.03ms<br/>95th percentile: 13.83ms ± 0.04ms<br/>99th percentile: 14.77ms ± 0.02ms<br/>Max time: 16.62ms ± 1.53ms |
| e2e_trip_ch | Ops: 98.23 ± 0.28 ops/s. Best: 97.54 ops/s<br/>Total: 10179.89ms ± 29.21ms. Best: 10153.17ms<br/>Min time: 1.52ms ± 0.14ms<br/>Mean time: 10.18ms ± 0.03ms<br/>Median time: 9.66ms ± 0.04ms<br/>95th percentile: 18.06ms ± 0.03ms<br/>99th percentile: 19.41ms ± 0.10ms<br/>Max time: 23.27ms ± 2.51ms | Ops: 98.05 ± 0.61 ops/s. Best: 96.54 ops/s<br/>Total: 10200.02ms ± 63.44ms. Best: 10145.88ms<br/>Min time: 1.52ms ± 0.15ms<br/>Mean time: 10.20ms ± 0.06ms<br/>Median time: 9.68ms ± 0.08ms<br/>95th percentile: 18.16ms ± 0.19ms<br/>99th percentile: 19.77ms ± 0.07ms<br/>Max time: 21.77ms ± 0.89ms |
| e2e_trip_mld | Ops: 58.95 ± 0.04 ops/s. Best: 58.91 ops/s<br/>Total: 16961.94ms ± 13.08ms. Best: 16929.92ms<br/>Min time: 1.56ms ± 0.19ms<br/>Mean time: 16.96ms ± 0.01ms<br/>Median time: 16.54ms ± 0.04ms<br/>95th percentile: 27.81ms ± 0.06ms<br/>99th percentile: 29.38ms ± 0.17ms<br/>Max time: 31.30ms ± 0.34ms | Ops: 58.73 ± 0.05 ops/s. Best: 58.66 ops/s<br/>Total: 17028.43ms ± 13.08ms. Best: 16999.59ms<br/>Min time: 1.66ms ± 0.20ms<br/>Mean time: 17.03ms ± 0.01ms<br/>Median time: 16.59ms ± 0.03ms<br/>95th percentile: 28.01ms ± 0.05ms<br/>99th percentile: 29.80ms ± 0.08ms<br/>Max time: 31.96ms ± 0.53ms |
| json-render | String: 6.83732ms<br/>Stringstream: 10.2285ms<br/>Vector: 6.60284ms | String: 6.74548ms<br/>Stringstream: 10.4877ms<br/>Vector: 7.10659ms |
| match_ch | Default radius: <br/>4.60339ms/req at 82 coordinate<br/>0.0561389ms/coordinate<br/>Radius 10m: <br/>16.1992ms/req at 82 coordinate<br/>0.197551ms/coordinate | Default radius: <br/>4.57027ms/req at 82 coordinate<br/>0.055735ms/coordinate<br/>Radius 10m: <br/>15.9323ms/req at 82 coordinate<br/>0.194296ms/coordinate |
| match_mld | Default radius: <br/>3.33234ms/req at 82 coordinate<br/>0.0406383ms/coordinate<br/>Radius 10m: <br/>11.563ms/req at 82 coordinate<br/>0.141012ms/coordinate | Default radius: <br/>3.18202ms/req at 82 coordinate<br/>0.0388052ms/coordinate<br/>Radius 10m: <br/>12.1316ms/req at 82 coordinate<br/>0.147946ms/coordinate |
| osrm_contract | Time: 94.34s Peak RAM: 195.88MB | Time: 94.53s Peak RAM: 195.90MB |
| osrm_customize | Time: 1.39s Peak RAM: 116.66MB | Time: 1.33s Peak RAM: 116.80MB |
| osrm_extract | Time: 11.91s Peak RAM: 409.95MB | Time: 11.96s Peak RAM: 411.39MB |
| osrm_partition | Time: 1.99s Peak RAM: 139.03MB | Time: 1.99s Peak RAM: 134.10MB |
| packedvector | random write:<br/>std::vector 11190 ms<br/>util::packed_vector 73919.9 ms<br/>slowdown: 6.60588<br/>random read:<br/>std::vector 11115.7 ms<br/>util::packed_vector 30600 ms<br/>slowdown: 2.75286 | random write:<br/>std::vector 11242.8 ms<br/>util::packed_vector 81998.9 ms<br/>slowdown: 7.29345<br/>random read:<br/>std::vector 11103.3 ms<br/>util::packed_vector 33758.1 ms<br/>slowdown: 3.04037 |
| random_match_ch | 500 matches, default radius<br/>ops: 220.16 ± 0.71 ops/s. best: 218.48ops/s.<br/>total: 258.91 ± 0.84ms. best: 258.10ms.<br/>avg: 4.54 ± 0.01ms<br/>min: 0.15 ± 0.01ms<br/>max: 25.77 ± 0.07ms<br/>p99: 25.77 ± 0.07ms<br/><br/>500 matches, radius=10<br/>ops: 63.50 ± 0.14 ops/s. best: 63.25ops/s.<br/>total: 1007.93 ± 2.29ms. best: 1004.56ms.<br/>avg: 15.75 ± 0.04ms<br/>min: 0.16 ± 0.00ms<br/>max: 251.54 ± 0.99ms<br/>p99: 251.54 ± 0.99ms<br/><br/>500 matches, radius=20<br/>ops: 15.08 ± 0.03 ops/s. best: 15.04ops/s.<br/>total: 4309.28 ± 7.92ms. best: 4297.06ms.<br/>avg: 66.30 ± 0.12ms<br/>min: 0.32 ± 0.00ms<br/>max: 1263.96 ± 5.91ms<br/>p99: 1263.96 ± 5.91ms | 500 matches, default radius<br/>ops: 223.20 ± 0.75 ops/s. best: 221.48ops/s.<br/>total: 255.37 ± 0.86ms. best: 254.50ms.<br/>avg: 4.48 ± 0.02ms<br/>min: 0.15 ± 0.01ms<br/>max: 25.52 ± 0.06ms<br/>p99: 25.52 ± 0.06ms<br/><br/>500 matches, radius=10<br/>ops: 64.06 ± 0.05 ops/s. best: 63.99ops/s.<br/>total: 999.00 ± 0.83ms. best: 997.27ms.<br/>avg: 15.61 ± 0.01ms<br/>min: 0.16 ± 0.00ms<br/>max: 249.75 ± 0.69ms<br/>p99: 249.75 ± 0.69ms<br/><br/>500 matches, radius=20<br/>ops: 15.19 ± 0.03 ops/s. best: 15.14ops/s.<br/>total: 4278.55 ± 7.12ms. best: 4270.46ms.<br/>avg: 65.82 ± 0.11ms<br/>min: 0.32 ± 0.00ms<br/>max: 1259.97 ± 3.75ms<br/>p99: 1259.97 ± 3.75ms |
| random_match_mld | 500 matches, default radius<br/>ops: 300.42 ± 1.47 ops/s. best: 296.95ops/s.<br/>total: 189.74 ± 0.93ms. best: 188.84ms.<br/>avg: 3.33 ± 0.02ms<br/>min: 0.13 ± 0.01ms<br/>max: 19.69 ± 0.02ms<br/>p99: 19.69 ± 0.02ms<br/><br/>500 matches, radius=10<br/>ops: 105.97 ± 0.15 ops/s. best: 105.73ops/s.<br/>total: 603.97 ± 0.87ms. best: 602.38ms.<br/>avg: 9.44 ± 0.01ms<br/>min: 0.15 ± 0.00ms<br/>max: 114.79 ± 0.41ms<br/>p99: 114.79 ± 0.41ms<br/><br/>500 matches, radius=20<br/>ops: 21.07 ± 0.07 ops/s. best: 20.96ops/s.<br/>total: 3084.70 ± 9.97ms. best: 3071.66ms.<br/>avg: 47.46 ± 0.15ms<br/>min: 0.20 ± 0.00ms<br/>max: 605.52 ± 2.00ms<br/>p99: 605.52 ± 2.00ms | 500 matches, default radius<br/>ops: 302.23 ± 1.93 ops/s. best: 297.67ops/s.<br/>total: 188.61 ± 1.21ms. best: 187.50ms.<br/>avg: 3.31 ± 0.02ms<br/>min: 0.13 ± 0.00ms<br/>max: 19.56 ± 0.07ms<br/>p99: 19.56 ± 0.07ms<br/><br/>500 matches, radius=10<br/>ops: 106.09 ± 0.07 ops/s. best: 106.00ops/s.<br/>total: 603.26 ± 0.39ms. best: 602.66ms.<br/>avg: 9.43 ± 0.01ms<br/>min: 0.15 ± 0.00ms<br/>max: 114.69 ± 0.42ms<br/>p99: 114.69 ± 0.42ms<br/><br/>500 matches, radius=20<br/>ops: 21.15 ± 0.04 ops/s. best: 21.10ops/s.<br/>total: 3072.72 ± 6.52ms. best: 3062.96ms.<br/>avg: 47.27 ± 0.10ms<br/>min: 0.20 ± 0.00ms<br/>max: 604.10 ± 2.62ms<br/>p99: 604.10 ± 2.62ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 22602.79 ± 45.44 ops/s. best: 22486.25ops/s.<br/>total: 442.43 ± 0.89ms. best: 441.76ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17144.99 ± 10.85 ops/s. best: 17132.85ops/s.<br/>total: 583.26 ± 0.37ms. best: 582.55ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.01ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13906.74 ± 13.88 ops/s. best: 13878.53ops/s.<br/>total: 719.08 ± 0.76ms. best: 718.09ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.20 ± 0.04ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 23015.40 ± 51.77 ops/s. best: 22944.39ops/s.<br/>total: 434.49 ± 0.98ms. best: 433.03ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17397.11 ± 24.91 ops/s. best: 17338.50ops/s.<br/>total: 574.81 ± 0.83ms. best: 573.94ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14037.66 ± 23.90 ops/s. best: 13996.29ops/s.<br/>total: 712.37 ± 1.21ms. best: 711.14ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 22697.61 ± 52.97 ops/s. best: 22573.68ops/s.<br/>total: 440.58 ± 1.03ms. best: 439.60ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17166.64 ± 25.70 ops/s. best: 17128.19ops/s.<br/>total: 582.53 ± 0.87ms. best: 581.24ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13891.25 ± 17.53 ops/s. best: 13852.42ops/s.<br/>total: 719.88 ± 0.92ms. best: 718.86ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 23232.01 ± 59.89 ops/s. best: 23086.13ops/s.<br/>total: 430.44 ± 1.11ms. best: 429.52ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.10 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17468.61 ± 16.80 ops/s. best: 17432.06ops/s.<br/>total: 572.46 ± 0.55ms. best: 571.90ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.01ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14089.80 ± 19.18 ops/s. best: 14054.67ops/s.<br/>total: 709.73 ± 0.97ms. best: 708.22ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.00ms<br/>p99: 0.13 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 510.96 ± 2.07 ops/s. best: 508.31ops/s.<br/>total: 1925.84 ± 7.78ms. best: 1913.93ms.<br/>avg: 1.96 ± 0.01ms<br/>min: 0.35 ± 0.00ms<br/>max: 3.36 ± 0.13ms<br/>p99: 2.85 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 598.80 ± 5.28 ops/s. best: 587.05ops/s.<br/>total: 1670.19 ± 15.66ms. best: 1653.85ms.<br/>avg: 1.67 ± 0.02ms<br/>min: 0.06 ± 0.00ms<br/>max: 3.96 ± 0.06ms<br/>p99: 3.45 ± 0.04ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1012.87 ± 4.51 ops/s. best: 1003.79ops/s.<br/>total: 971.53 ± 4.35ms. best: 966.18ms.<br/>avg: 0.99 ± 0.00ms<br/>min: 0.23 ± 0.00ms<br/>max: 1.56 ± 0.03ms<br/>p99: 1.37 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1145.07 ± 6.47 ops/s. best: 1132.43ops/s.<br/>total: 873.35 ± 4.95ms. best: 866.56ms.<br/>avg: 0.87 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.95 ± 0.02ms<br/>p99: 1.76 ± 0.03ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 512.66 ± 1.27 ops/s. best: 511.18ops/s.<br/>total: 1919.42 ± 4.76ms. best: 1912.64ms.<br/>avg: 1.95 ± 0.00ms<br/>min: 0.34 ± 0.00ms<br/>max: 3.48 ± 0.16ms<br/>p99: 2.85 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 604.08 ± 1.00 ops/s. best: 602.72ops/s.<br/>total: 1655.43 ± 2.74ms. best: 1650.06ms.<br/>avg: 1.66 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 3.88 ± 0.02ms<br/>p99: 3.39 ± 0.01ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 1035.47 ± 4.76 ops/s. best: 1028.05ops/s.<br/>total: 950.32 ± 4.38ms. best: 945.76ms.<br/>avg: 0.97 ± 0.00ms<br/>min: 0.23 ± 0.00ms<br/>max: 1.54 ± 0.02ms<br/>p99: 1.34 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1171.73 ± 5.30 ops/s. best: 1161.07ops/s.<br/>total: 853.46 ± 3.87ms. best: 848.13ms.<br/>avg: 0.85 ± 0.00ms<br/>min: 0.04 ± 0.00ms<br/>max: 2.96 ± 0.03ms<br/>p99: 1.73 ± 0.02ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 250.55 ± 0.76 ops/s. best: 249.34ops/s.<br/>total: 3927.42 ± 12.11ms. best: 3905.09ms.<br/>avg: 3.99 ± 0.01ms<br/>min: 0.35 ± 0.01ms<br/>max: 8.55 ± 0.06ms<br/>p99: 6.62 ± 0.07ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 245.45 ± 0.24 ops/s. best: 244.93ops/s.<br/>total: 4074.11 ± 4.02ms. best: 4069.75ms.<br/>avg: 4.07 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 9.56 ± 0.70ms<br/>p99: 8.14 ± 0.07ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 333.77 ± 0.62 ops/s. best: 332.84ops/s.<br/>total: 2948.19 ± 5.46ms. best: 2938.22ms.<br/>avg: 3.00 ± 0.01ms<br/>min: 0.30 ± 0.00ms<br/>max: 7.32 ± 0.30ms<br/>p99: 5.13 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 305.02 ± 1.65 ops/s. best: 301.39ops/s.<br/>total: 3278.56 ± 17.85ms. best: 3257.82ms.<br/>avg: 3.28 ± 0.02ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.52 ± 0.36ms<br/>p99: 6.40 ± 0.05ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 251.67 ± 0.64 ops/s. best: 250.66ops/s.<br/>total: 3909.85 ± 10.16ms. best: 3897.78ms.<br/>avg: 3.97 ± 0.01ms<br/>min: 0.34 ± 0.00ms<br/>max: 8.69 ± 0.10ms<br/>p99: 6.59 ± 0.04ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 244.48 ± 0.22 ops/s. best: 244.17ops/s.<br/>total: 4090.28 ± 3.69ms. best: 4083.71ms.<br/>avg: 4.09 ± 0.00ms<br/>min: 0.05 ± 0.00ms<br/>max: 9.13 ± 0.27ms<br/>p99: 8.14 ± 0.03ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 334.94 ± 1.12 ops/s. best: 333.78ops/s.<br/>total: 2937.91 ± 9.82ms. best: 2921.31ms.<br/>avg: 2.99 ± 0.01ms<br/>min: 0.29 ± 0.00ms<br/>max: 7.29 ± 0.29ms<br/>p99: 5.17 ± 0.03ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 305.14 ± 0.63 ops/s. best: 303.95ops/s.<br/>total: 3277.18 ± 6.81ms. best: 3268.38ms.<br/>avg: 3.28 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.21 ± 0.12ms<br/>p99: 6.37 ± 0.05ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1509.87 ± 10.36 ops/s. best: 1486.46ops/s.<br/>total: 165.59 ± 1.15ms. best: 164.59ms.<br/>avg: 0.66 ± 0.00ms<br/>min: 0.47 ± 0.00ms<br/>max: 1.00 ± 0.26ms<br/>p99: 0.82 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 178.10 ± 0.22 ops/s. best: 177.75ops/s.<br/>total: 1403.69 ± 1.72ms. best: 1400.76ms.<br/>avg: 5.61 ± 0.01ms<br/>min: 5.06 ± 0.00ms<br/>max: 6.45 ± 0.49ms<br/>p99: 6.07 ± 0.04ms<br/><br/>250 tables, 50 coordinates<br/>ops: 88.03 ± 0.11 ops/s. best: 87.89ops/s.<br/>total: 2839.82 ± 3.45ms. best: 2833.05ms.<br/>avg: 11.36 ± 0.01ms<br/>min: 10.61 ± 0.03ms<br/>max: 12.87 ± 0.89ms<br/>p99: 12.06 ± 0.07ms | 250 tables, 3 coordinates<br/>ops: 1532.94 ± 11.79 ops/s. best: 1503.05ops/s.<br/>total: 163.10 ± 1.27ms. best: 162.03ms.<br/>avg: 0.65 ± 0.01ms<br/>min: 0.46 ± 0.00ms<br/>max: 0.99 ± 0.26ms<br/>p99: 0.80 ± 0.02ms<br/><br/>250 tables, 25 coordinates<br/>ops: 180.14 ± 0.15 ops/s. best: 179.94ops/s.<br/>total: 1387.79 ± 1.14ms. best: 1386.13ms.<br/>avg: 5.55 ± 0.00ms<br/>min: 5.00 ± 0.01ms<br/>max: 6.02 ± 0.02ms<br/>p99: 5.96 ± 0.02ms<br/><br/>250 tables, 50 coordinates<br/>ops: 88.70 ± 0.06 ops/s. best: 88.63ops/s.<br/>total: 2818.42 ± 1.86ms. best: 2814.61ms.<br/>avg: 11.27 ± 0.01ms<br/>min: 10.51 ± 0.01ms<br/>max: 12.37 ± 0.12ms<br/>p99: 12.09 ± 0.05ms |
…