@DennisOSRM 2024-06-21
I think my main concern here is that with a priority queue the type itself already tells a lot about its usage. 

Using a vector with the right heap functions gives equivalent functionality, but you have to understand the code (and the implementation of std::priority_queue) to know that there is a priority ordering. 

I hope this makes more sense.