@SiarheiFedartsou 2024-06-21
> I think my main concern here is that with a priority queue the type itself already tells a lot about its usage.
> 
> Using a vector with the right heap functions gives equivalent functionality, but you have to understand the code (and the implementation of std::priority_queue) to know that there is a priority ordering.
> 
> I hope this makes more sense.

Okay, got it now, will it be okay for you if we would introduce our own `PriorityQueue` class (with proper tests etc)? According to my local tests (which are not that noisy as our CI benchmarks for now :) ) this is a quite good optimisation, so I would like to eventually somehow have it in master.