@SiarheiFedartsou 2024-06-21
@DennisOSRM WDYT about updated implementation? 😀